+++ backend/src/main/java/ca/bc/gov/restapi/results/oracle/entity/cover/ForestCoverDetailedSpeciesProjection.java
@@ -6,6 +6,6 @@ public interface ForestCoverDetailedSpeciesProjection extends ForestCoverSpecies
 
   Long getAverageAge();
 
-  Long getAverageHeight();
+  Float getAverageHeight();
 
 }
